+++ lib/view_component/base.rb
@@ -40,6 +40,23 @@ module ViewComponent
       # noop
     end
 
+    # @!macro [attach] deprecated_generate_mattr_accessor
+    #   @method generate_$1
+    #   @deprecated Use `#generate.$1` instead. Will be removed in v3.0.0.
+    def self._deprecated_generate_mattr_accessor(name)
+      define_singleton_method("generate_#{name}".to_sym) do
+        generate.public_send(name)
+      end
+      define_singleton_method("generate_#{name}=".to_sym) do |value|
+        generate.public_send("#{name}=".to_sym, value)
+      end
+    end
+
+    _deprecated_generate_mattr_accessor :distinct_locale_files
+    _deprecated_generate_mattr_accessor :locale
+    _deprecated_generate_mattr_accessor :sidecar
+    _deprecated_generate_mattr_accessor :stimulus_controller
+
     # Entrypoint for rendering components.
     #
     # - `view_context`: ActionView context from calling view
@@ -271,33 +288,6 @@ module ViewComponent
     #
     mattr_accessor :render_monkey_patch_enabled, instance_writer: false, default: true
 
-    # Always generate a Stimulus controller alongside the component:
-    #
-    #     config.view_component.generate_stimulus_controller = true
-    #
-    # Defaults to `false`.
-    #
-    mattr_accessor :generate_stimulus_controller, instance_writer: false, default: false
-
-    # Always generate translations file alongside the component:
-    #
-    #     config.view_component.generate_locale = true
-    #
-    # Defaults to `false`.
-    #
-    mattr_accessor :generate_locale, instance_writer: false, default: false
-
-    # Always generate as many translations files as available locales:
-    #
-    #     config.view_component.generate_distinct_locale_files = true
-    #
-    # Defaults to `false`.
-    #
-    # One file will be generated for each configured `I18n.available_locales`.
-    # Fallback on `[:en]` when no available_locales is defined.
-    #
-    mattr_accessor :generate_distinct_locale_files, instance_writer: false, default: false
-
     # Path for component files
     #
     #     config.view_component.view_component_path = "app/my_components"
@@ -310,17 +300,44 @@ module ViewComponent
     #
     #     config.view_component.component_parent_class = "MyBaseComponent"
     #
-    # Defaults to "ApplicationComponent" if defined, "ViewComponent::Base" otherwise.
+    # Defaults to nil. If this is falsy, generators will use
+    # "ApplicationComponent" if defined, "ViewComponent::Base" otherwise.
     #
     mattr_accessor :component_parent_class, instance_writer: false
 
+    # Configuration for generators.
+    #
+    # All options under this namespace default to `false` unless otherwise
+    # stated.
+    #
+    # #### #sidecar
+    #
     # Always generate a component with a sidecar directory:
     #
-    #     config.view_component.generate_sidecar = true
+    #     config.view_component.generate.sidecar = true
     #
-    # Defaults to `false`.
+    # #### #stimulus_controller
     #
-    mattr_accessor :generate_sidecar, instance_writer: false, default: false
+    # Always generate a Stimulus controller alongside the component:
+    #
+    #     config.view_component.generate.stimulus_controller = true
+    #
+    # #### #locale
+    #
+    # Always generate translations file alongside the component:
+    #
+    #     config.view_component.generate.locale = true
+    #
+    # #### #distinct_locale_files
+    #
+    # Always generate as many translations files as available locales:
+    #
+    #     config.view_component.generate.distinct_locale_files = true
+    #
+    # One file will be generated for each configured `I18n.available_locales`,
+    # falling back to `[:en]` when no `available_locales` is defined.
+    #
+    mattr_accessor :generate, instance_writer: false, default: ActiveSupport::OrderedOptions.new(false)
 
     class << self
       # @private